Separate pantry function
